requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
einops==0.7.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
Flask==3.0.2
fonttools==4.50.0
freetype-py==2.4.0
fsspec==2023.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
idna==3.4
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mesh-intersection==0.1.0
mpmath==1.3.0
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.0
numpy==1.22.0
oauthlib==3.2.2
open3d==0.18.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyglet==2.0.15
Pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.28.1
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tqdm==4.66.2
traitlets==5.14.2
trimesh==4.2.4
triton==2.1.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.2
widgetsnbextension==4.0.10
zipp==3.18.1
